+++ CHANGELOG.md
@@ -67,6 +67,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `ngl_config.{width,height,samples}` and `ngl_resize()` arguments are now unsigned
 
 ### Removed
+- `DrawPath.aspect_ratio`, it now matches the scene aspect ratio
 - `Text.aspect_ratio`, it now matches the viewport aspect ratio
 - `GraphicConfig.scissor_test` parameter
 - Support for Android < 9.0